requirements.txt
bs4
requests
PyPDF2
langchain
sentence-transformers
faiss-cpu
google-generativeai
fastapi
uvicorn
BaseModel
python-dotenv

